Fix: 필터링 북마크리스트dto 반환 로직 추가

+++ server/itclips/src/main/java/com/ssafy/itclips/recommend/dto/SimilarBookmarkResponse.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 public class SimilarBookmarkResponse {
 
     @JsonProperty("bookmark_list_id")
-    private Integer bookmarkListId;
+    private Long bookmarkListId;
 
     @JsonProperty("similarity_score")
     private Double similarityScore;
